requirements.txt
altgraph==0.17.4
annotated-types==0.7.0
attrs==25.3.0
certifi==2025.8.3
charset-normalizer==3.4.2
idna==3.10
isodate==0.7.2
lxml==6.0.0
numpy==2.2.6
onvif_zeep==0.2.12
opencv-python==4.12.0.88
packaging==25.0
pefile==2023.2.7
platformdirs==4.3.8
pydantic==2.11.7
pydantic_core==2.33.2
pyinstaller==6.14.2
pyinstaller-hooks-contrib==2025.8
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.17.0
PyQtChart==5.15.7
PyQtChart-Qt5==5.15.2
pytz==2025.2
pywin32-ctypes==0.2.3
PyYAML==6.0.2
requests==2.32.4
requests-file==2.1.0
requests-toolbelt==1.0.0
ruamel.yaml==0.18.14
ruamel.yaml.clib==0.2.12
setuptools==80.9.0
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
zeep==4.3.1
